requirements.txt
PyYAML==3.12
Flask==0.11.1
Flask-Menu
Flask-Session
